backend/requirements.txt
fastapi
uvicorn[standard]
finance-datareader
pandas
google-generativeai
supabase
python-dotenv
pydantic-settings
slowapi
httpx
feedparser
yfinance
pykrx
multipledispatch
beautifulsoup4
lxml
pypdf
OpenDartReader
openpyxl
